Clear AddComponent function (#75)
* Clear code Breakout

* Clear code Snake

+++ Breakout/Script/GameLevel.h
@@ -11,10 +11,10 @@ class GameLevel
 {
 private:
     const std::string fileName;
-    unsigned int levelHeight;
-    unsigned int levelWidth;
+    size_t levelHeight;
+    size_t levelWidth;
 
-    void Init(const std::vector<std::vector<unsigned int>>& brickData, bool areBricksActive);
+    void Init(const std::vector<std::vector<size_t>>& brickData, bool areBricksActive);
     void CalcBrickSize(const Learning2DEngine::Render::Resolution& resolution);
 public:
     std::vector<BrickController*> bricks;
